rate-service-refactoring: move out repos
