+++ src/services/rates/RateEstimator.ts
@@ -1,25 +1,22 @@
 import { Task, of as taskOf } from "folktale/concurrency/task";
 import { Maybe, of as maybeOf } from 'folktale/maybe';
 
-import { BigNumber } from "@waves/data-entities";
 import {
   identity,
-  uniqWith,
   map,
   chain,
-  partition,
 } from 'ramda';
 
 import { tap } from "../../utils/tap";
-import * as LRU from 'lru-cache';
 import { AssetIdsPair, RateInfo } from "../../types";
 import { AppError, DbError } from "../../errorHandling";
 import { PgDriver } from "db/driver";
 import * as knex from 'knex'
 
 import makeSql from './sql';
-import { RateCache, RateInfoLookup, RateCacheKey } from './repo';
-import { WavesId, pairIsSymmetric, flip } from './data';
+import { partitionByPreCount } from './repo';
+import  RateCache, { RateCacheKey } from './repo/impl/RateCache';
+import RateInfoLookup from './repo/impl/RateInfoLookup'
 import { maybeIsNone } from "./util";
 
 const pg = knex({ client: 'pg' });
@@ -29,88 +26,11 @@ type ReqAndRes<TReq, TRes> = {
   res: Maybe<TRes>
 }
 
-function requestData(pair: AssetIdsPair): AssetIdsPair[] {
-  if (pair.amountAsset === WavesId || pair.priceAsset === WavesId) {
-    return [pair, flip(pair)]
-  }
-
-  const wavesL: AssetIdsPair = {
-    amountAsset: pair.amountAsset,
-    priceAsset: WavesId
-  }
-
-  const wavesR: AssetIdsPair = {
-    amountAsset: pair.priceAsset,
-    priceAsset: WavesId
-  }
-
-  return [
-    pair, flip(pair),
-    wavesL, flip(wavesL),
-    wavesR, flip(wavesR)
-  ]
-}
-
-const pairsEq = (pair1: AssetIdsPair, pair2: AssetIdsPair): boolean =>
-  pair1.amountAsset === pair2.amountAsset && pair1.priceAsset === pair2.priceAsset
-
-type PairsForRequest = {
-  preCount: RateInfo[],
-  toBeRequested: AssetIdsPair[]
-}
-
-const partitionByPreCount = (cache: RateCache, pairs: AssetIdsPair[], getCacheKey: (pair: AssetIdsPair) => Maybe<RateCacheKey>): PairsForRequest => {
-  const [eq, uneq] = partition(pairIsSymmetric, pairs)
-
-  const allPairsToRequest = uniqWith(
-    pairsEq,
-    chain(it => requestData(it), uneq)
-  )
-
-  const [cached, uncached] = partition(
-    it => cache.has(
-      getCacheKey(it)
-    ),
-    allPairsToRequest
-  )
-
-  const cachedRates: RateInfo[] = cached.map(
-    pair => (
-      {
-        amountAsset: pair.amountAsset,
-        priceAsset: pair.priceAsset,
-        current: cache.get(
-          getCacheKey(pair)
-        ).getOrElse(new BigNumber(0))
-      }
-    )
-  )
-
-  const eqRates: RateInfo[] = eq.map(
-    (pair) => (
-      {
-        current: new BigNumber(1),
-          ...pair
-      }
-    )
-  )
-
-  return {
-    preCount: cachedRates.concat(eqRates),
-    toBeRequested: uncached
-  }
-}
-
 export default class RateEstimator {
-  private readonly cache: RateCache
-  
   constructor(
-    lru: LRU<string, BigNumber>,
+    private readonly cache: RateCache,
     private readonly pgp: PgDriver
-  ) {
-
-    this.cache = new RateCache(lru)
-  }
+  ) { }
 
   estimate(pairs: AssetIdsPair[], matcher: string, timestamp: Maybe<Date>): Task<AppError, ReqAndRes<AssetIdsPair, RateInfo>[]> {
     const shouldCache = maybeIsNone(timestamp)
