+++ src/services/rates/data.ts
@@ -16,3 +16,28 @@ export function flip(pair: AssetIdsPair): AssetIdsPair {
     priceAsset: pair.amountAsset
   }
 }
+
+export const pairsEq = (pair1: AssetIdsPair, pair2: AssetIdsPair): boolean =>
+  pair1.amountAsset === pair2.amountAsset && pair1.priceAsset === pair2.priceAsset
+
+export function generatePossibleRequestItems(pair: AssetIdsPair): AssetIdsPair[] {
+  if (pair.amountAsset === WavesId || pair.priceAsset === WavesId) {
+    return [pair, flip(pair)]
+  }
+
+  const wavesL: AssetIdsPair = {
+    amountAsset: pair.amountAsset,
+    priceAsset: WavesId
+  }
+
+  const wavesR: AssetIdsPair = {
+    amountAsset: pair.priceAsset,
+    priceAsset: WavesId
+  }
+
+  return [
+    pair, flip(pair),
+    wavesL, flip(wavesL),
+    wavesR, flip(wavesR)
+  ]
+}
